+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Compatibility:
 * Implemented `$LOAD_PATH.resolve_feature_path`.
 * Add `Pathname#/` alias to `Pathname#+` (#2178).
 * Fixed issue with large `Integer`s in `Math.log` (#2184).
+* Updated `Regexp.last_match` to support `Symbol` and `String` parameter (#2179).
 
 Performance: